+++ HappyHomeDesigner/HappyHomeDesigner/stage.h
@@ -16,7 +16,7 @@
 //*****************************************************************************
 #define START_STAGE		(STAGE_RESULT)	// 開始時のステージ
 #define STR(var) #var					// 引数にした変数を変数名を示す文字列リテラルとして返すマクロ関数
-#define START_PRICE		(0)				//初期金額
+#define START_PRICE		((long long)0)				//初期金額
 
 /*******************************************************************************
 * 構造体定義
@@ -60,8 +60,8 @@ void		DrawStage(void);
 
 void		SetStage(int nStage);
 int			GetStage(void);
-int			GetPrice(void);
-void		AddPrice(int price);
+long long	GetPrice(void);
+void		AddPrice(long long price);
 void		SetGameEnd(int no);
 int			GetGameEnd(void);
